requirements.txt
skydive-client>=0.4.2
